+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 * Initial implementation of Record.Checksum(), intended for easy and fast comparison of recordable objects.
 * Initial implementation of Record.ChecksumDiff(), intended for easy and fast detection of how recordable objects differ.
 * Added Intent property to Recorder to indicate Serialization/Cloning/Checksum purpose.
+* Official support for nullable value types.
 
 ### Improved
 * Exception messages for ConfigErrors() and PostLoad() now include the Dec name.